Escape dynamic values by default
Values of the dict returned by a renderer function decorated with
`@renders` will be escaped by default. To unescape them, use `raw()`
or manually cast them into bytes.

This will impact performance. People comparing it with other
templating libraries, need to check if those libraries auto
escapes the variable (i.e. with safety turned on).

This change is going to break a lot of things as I am rewriting the original
function instead of creating something like `@safely_render`. I am
taking this risk since github is showing that I am the only
(open source) user of this lib and this is still in beta.